+++ R/personnel.formatAC.R
@@ -23,7 +23,11 @@ personnel.formatAC <- function(
 {
   
   #### Données de référence ####
-  TpsWOT <- structure(list(id = integer(0), tpswot_date = structure(numeric(0), class = "Date"), 
+  annee_ac <- str_replace(projet, "Convention-cadre AERMC-FD39-", "")
+  date_debut_ac <- glue("{annee_ac}-01-01")
+  date_fin_ac <- glue("{annee_ac}-12-31")
+  
+  tps_w_ot <- structure(list(id = integer(0), tpswot_date = structure(numeric(0), class = "Date"), 
                            tpswot_domaine = character(0), tpswot_activite = character(0), 
                            tpswot_sous_projet = character(0), tpswot_personnel = integer(0), 
                            tpswot_heuredebut = character(0), tpswot_heurefin = character(0), 
@@ -33,31 +37,31 @@ personnel.formatAC <- function(
                            `_modif_date` = structure(numeric(0), tzone = "", class = c("POSIXct", 
                                                                                        "POSIXt"))), row.names = integer(0), class = c("tbl_df", 
                                                                                                                                       "tbl", "data.frame"))
-  Postes <- tbl(dbD, dbplyr::in_schema("fd_referentiels", "gestion_postes")) %>% select(id, gestpost_poste_libelle) %>% collect(n = Inf) %>% rename(idposte = id)
-  Personnels <- tbl(dbD, dbplyr::in_schema("fd_referentiels", "gestion_operateurs")) %>% filter(gestop_mo == 3 & gestop_type == "Salarié") %>% filter(is.na(gestop_datefinactivite)) %>% select(id:gestop_qualite) %>% collect(n = Inf) %>% left_join(Postes, by = c("gestop_qualite" = "gestpost_poste_libelle")) %>% mutate(gestop_qualite = idposte) %>% select(-idposte)
-  if(Personnels %>% filter(is.na(gestop_qualite)) %>% nrow() != 0) stop("Présence de personnels sans poste de travail clairement défini")
+  postes <- tbl(dbD, dbplyr::in_schema("fd_referentiels", "gestion_postes")) %>% select(id, gestpost_poste_libelle) %>% collect(n = Inf) %>% rename(idposte = id)
+  personnels <- tbl(dbD, dbplyr::in_schema("fd_referentiels", "gestion_operateurs")) %>% filter(gestop_mo == 3 & gestop_type == "Salarié") %>% filter(is.na(gestop_datefinactivite)|(gestop_datefinactivite >= date_debut_ac & gestop_datefinactivite <= date_fin_ac)) %>% filter(!(gestop_prenom %in% c("Collectif", "recruter", "Admin"))) %>% select(id:gestop_qualite) %>% collect(n = Inf) %>% left_join(postes, by = c("gestop_qualite" = "gestpost_poste_libelle")) %>% mutate(gestop_qualite = idposte) %>% select(-idposte)
+  if(personnels %>% filter(is.na(gestop_qualite)) %>% nrow() != 0) stop("Présence de personnels sans poste de travail clairement défini")
   
   if(is.na(projet)) stop("Pas de projet spécifié")
-  if(is.numeric(projet)){Projets <- tbl(dbD, dbplyr::in_schema("fd_production", "projets_liste")) %>% filter(id == projet) %>% collect(n = Inf)} # Si le projet est recherché via son id
-  if(!is.numeric(projet)){Projets <- tbl(dbD, dbplyr::in_schema("fd_production", "projets_liste")) %>% filter(prjlst_projet == projet) %>% collect(n = Inf)} # Si le projet est recherché via son intitulé
+  if(is.numeric(projet)){projets <- tbl(dbD, dbplyr::in_schema("fd_production", "projets_liste")) %>% filter(id == projet) %>% collect(n = Inf)} # Si le projet est recherché via son id
+  if(!is.numeric(projet)){projets <- tbl(dbD, dbplyr::in_schema("fd_production", "projets_liste")) %>% filter(prjlst_projet == projet) %>% collect(n = Inf)} # Si le projet est recherché via son intitulé
   
   #### Tests ####
-  if(grepl("Convention-cadre AERMC-FD39-", Projets$prjlst_projet) == FALSE) stop("Le projet spécifié n'est pas de la forme Convention-cadre AERMC-FD39-")
-  if(all(colnames(data) %in% colnames(TpsWOT)) == FALSE) stop("Données en entrée pas au format de la table tpstravail_opentime")
+  if(grepl("Convention-cadre AERMC-FD39-", projets$prjlst_projet) == FALSE) stop("Le projet spécifié n'est pas de la forme Convention-cadre AERMC-FD39-")
+  if(all(colnames(data) %in% colnames(tps_w_ot)) == FALSE) stop("Données en entrée pas au format de la table tpstravail_opentime")
   
   #### Réencodage ####
-  projet_id <- Projets %>% select(id) %>% pull()
-  anneeAC <- str_replace(Projets$prjlst_projet, "Convention-cadre AERMC-FD39-", "")
+  projet_id <- projets %>% select(id) %>% pull()
+  annee_ac <- str_replace(projets$prjlst_projet, "Convention-cadre AERMC-FD39-", "")
   pattern_1 <- c("^AC AE")
-  pattern_2 <- c(glue("{Projets$id} - AC AE"))
-  data <-
+  pattern_2 <- c(glue("{projets$id} - AC AE"))
+  data_v2 <-
     data %>% 
-    filter(year(tpswot_date) == anneeAC) %>% 
+    filter(year(tpswot_date) == annee_ac) %>% 
     filter(tpswot_domaine == "Projets") %>% 
     filter(str_detect(tpswot_activite, pattern_1) | str_detect(tpswot_activite, pattern_2)) %>% 
     # filter() %>% 
     # union(data %>% 
-    #         filter(year(tpswot_date) == anneeAC) %>% 
+    #         filter(year(tpswot_date) == annee_ac) %>% 
     #         filter(tpswot_domaine == "Projets") %>% 
     #         filter(grepl("1.1|1.2|1.3|2\\.1|2\\.2|2.3|2.4|2.5|3.1|4.1|4.2", tpswot_sous_projet))
     # ) %>% 
@@ -78,24 +82,24 @@ personnel.formatAC <- function(
     mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "2\\.2"), 2.2, tpswdetail_sousactionaermc)) %>% # Il faut un échappement du point car sinon il cherche 2*2, soit par ex 2020
     mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "2\\.2"), 2.2, tpswdetail_sousactionaermc)) %>% # Il faut un échappement du point car sinon il cherche 2*2, soit par ex 2020
     mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "2.3"), 2.3, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "2.3"), 2.3, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "2.4"), 2.4, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "2.4"), 2.4, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "2.5"), 2.5, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "2.5"), 2.5, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "3.1"), 3.1, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "3.1"), 3.1, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "4.1"), 4.1, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "4.1"), 4.1, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "4.2"), 4.2, tpswdetail_sousactionaermc)) %>%
-    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "4.2"), 4.2, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "2.3"), 2.3, tpswdetail_sousactionaermc)) %>% # Il faut un échappement du point car sinon il cherche 2*2, soit par ex 2020
+    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "2\\.4"), 2.4, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "2\\.4"), 2.4, tpswdetail_sousactionaermc)) %>% # Il faut un échappement du point car sinon il cherche 2*4, soit par ex l'identifiant de projet 224
+    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "2\\.5"), 2.5, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "2\\.5"), 2.5, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "3\\.1"), 3.1, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "3\\.1"), 3.1, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "4\\.1"), 4.1, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "4\\.1"), 4.1, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(str_detect(tpswot_activite, "4\\.2"), 4.2, tpswdetail_sousactionaermc)) %>%
+    mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_sous_projet, "4\\.2"), 4.2, tpswdetail_sousactionaermc)) %>%
     mutate(tpswdetail_sousactionaermc = ifelse(is.na(tpswdetail_sousactionaermc) & str_detect(tpswot_activite, "AC AE 3.1"), 3.1, tpswdetail_sousactionaermc)) %>% # "AC AE 3.1 Hydrologie Lac des Rousses" mystérieusement non traité par phrase classique plus haut
     mutate(tpswdetail_actionaermc = str_sub(tpswdetail_sousactionaermc, 1, 1)) %>% # Déduit à partir de la sous-action
-    left_join(Personnels %>% mutate(tpswdetail_personnel = glue("{gestop_prenom} {gestop_nom}")), by = c("tpswot_personnel" = "id")) %>% 
+    left_join(personnels %>% mutate(tpswdetail_personnel = glue("{gestop_prenom} {gestop_nom}")), by = c("tpswot_personnel" = "id")) %>% 
     rename(tpswdetail_poste = gestop_qualite) %>% 
     select(-contains("gestop")) %>% 
     mutate(tpswdetail_statut = case_when(.$tpswdetail_poste %in% c(2, 3, 4, 5, 6, 7, 8) ~ "Expertise",
-                                         .$tpswdetail_poste %in% c(1, 9, 10) ~ "Surveillance")) %>% 
+                                         .$tpswdetail_poste %in% c(1, 9, 10, 13) ~ "Surveillance")) %>% 
     mutate(tpswdetail_detail = NA_character_) %>% 
     mutate(tpswdetail_detail = ifelse(is.na(.$tpswot_sous_projet) & is.na(.$tpswot_remarques), tpswot_activite, tpswdetail_detail)) %>% 
     mutate(tpswdetail_detail = ifelse(is.na(.$tpswot_remarques), glue('{tpswot_activite} - {tpswot_sous_projet}'), tpswdetail_detail)) %>% 
@@ -103,7 +107,7 @@ personnel.formatAC <- function(
     mutate(tpswdetail_detail = ifelse(is.na(.$tpswdetail_detail), glue('{tpswot_domaine} - {tpswot_activite} - {tpswot_remarques}'), tpswdetail_detail))
     
   #### Sortie des données ####
-  return(data)
+  return(data_v2)
   
 } # Fin de la fonction