+++ R/stations.territoire.R
@@ -131,6 +131,8 @@ stations.territoire <- function(
     # DBI::dbDisconnect(dbD)
     if(Liste == F){
       if(class(Territoire) != "character"){stop("Liste de territoires pas au bon format")}
+      if(grepl("BIE|Bienne", Territoire)){Territoire <- c("39.19", "39.62", "39.16", "39.17", "39.21", "39.24", "39.18", "39.22", "39.23", "39.18", "39.18", "39.25")}
+      if(length(Territoire) == 1) {if(grepl("ORB|Orbe", Territoire)){Territoire <- c("39.33")}}
     }
     if(Liste == T){
       #Territoire <- select.list(sort(contextesPDPG$hycont_contexte_libelle), multiple = T) # Liste directement dans R, avec sélection par numéro
@@ -158,19 +160,21 @@ stations.territoire <- function(
   
   #### GEMAPI ####
   if(Echelle == "GEMAPI"){
-    # stop("Traitement des entités GEMAPI à développer")
-    dbD <- BDD.ouverture("Data")
-    mo_territoires <- sf::st_read(dbD, query = "SELECT * FROM fd_referentiels.gestion_moterritoires WHERE (gestmoterr_activite IS TRUE);")
-    DBI::dbDisconnect(dbD)
-    if(Liste == F){
-      if(class(Territoire) != "character"){stop("Liste de territoires pas au bon format")}
-    }
-    if(Liste == T){
-      #Territoire <- select.list(sort(mo_territoires$gestmoterr_intitule), multiple = T) # Liste directement dans R, avec sélection par numéro
-      Territoire <- tcltk::tk_select.list(sort(mo_territoires$gestmoterr_intitule), multiple = T) # Pour avoir un menu qui s'ouvre
-    }
-    Sortie <- mo_territoires %>% mutate(nometcode = paste0(gestmoterr_intitule, " - ", id)) %>% filter(grepl(paste(Territoire,collapse="|"), nometcode)) %>% select(-nometcode)
-    if(nrow(Sortie) == 0){warning("Aucun résultat")}
+    stop("Traitement des entités GEMAPI à développer")
+    # Il faut créer une table avec les territoires d'action qui fait référence à la table des MO et qui historise (champ actif ou tri/filtre sur date d'ajout la plus récente ?) les versions des différents territoires d'action, avec un champ de sous-thème GEMAP/N2000 (cas du PNRHJ), avec une seule combinaison active possible
+    
+    # dbD <- BDD.ouverture("Data")
+    # communes <- sf::st_read(dbD, query = "SELECT * FROM fd_referentiels.topographie_communes WHERE (tpcomm_departement_insee = '39');")
+    # DBI::dbDisconnect(dbD)
+    # if(Liste == F){
+    #   if(class(Territoire) != "character"){stop("Liste de territoires pas au bon format")}
+    # }
+    # if(Liste == T){
+    #   #Territoire <- select.list(sort(communes$tpcomm_commune_libelle), multiple = T) # Liste directement dans R, avec sélection par numéro
+    #   Territoire <- tcltk::tk_select.list(sort(communes$tpcomm_commune_libelle), multiple = T) # Pour avoir un menu qui s'ouvre
+    # }
+    # Sortie <- communes %>% mutate(nometcode = paste0(tpcomm_commune_libelle, " - ", tpcomm_commune_insee)) %>% filter(grepl(paste(Territoire,collapse="|"), nometcode)) %>% select(-nometcode)
+    # if(nrow(Sortie) == 0){warning("Aucun résultat")}
   }
   
   #### MO ####